+++ tests/test_prg_engine_runtime_surface_functions_tests.h
@@ -610,4 +610,5 @@ namespace copperfin::runtime_surface_tests
     void test_getobject_reuses_existing_class_and_source_targets();
     void test_reportlistener_configuration_contracts_are_scoped_and_deterministic();
     void test_reportlistener_strict_configuration_uses_admitted_bytes();
+    void test_wvisible_tracks_modeled_vfp_windows();
 }